fast_mcp/requirements.txt
mcp
requests
google-generativeai
python-dotenv
